requirements.txt
requests>=2.32.0
pydantic>=2.8.0
jsonschema>=4.22.0
Pillow>=10.4.0
ollama>=0.5.3
unstructured>=0.14.0
unstructured[doc]>=0.18.1
docx2txt>=0.8
python-magic>=0.4.27
openai>=1.40.0
python-dotenv>=1.1.0
pymupdf>=1.24.2
python-pptx>=0.6.23
charset-normalizer>=3.3.2
faiss-cpu>=1.8.0
filelock>=3.14.0
fastapi>=0.115.0
uvicorn>=0.34.0
jinja2>=3.1.2
python-multipart>=0.0.12
watchdog>=6.0.0
httpx>=0.28.1
rich>=13.10.0
rapidfuzz>=6.0.0
tldextract>=3.9.0